pack read and write

+++ include/my_io.h
@@ -1,9 +1,9 @@
 #ifndef MY_IO_H
 #define MY_IO_H
 
-ssize_t readn(int fd, void * data, int len);
+int readn(int fd, void * data, int len);
 
-ssize_t writen(int fd, const void *data, size_t n);
+int writen(int fd, const void *data, size_t n);
 
 
 #endif
